refactor(attendance): move Current view to /attendance/current
The Current tab shared the bare /attendance URL, so the index route and
the first tab were the same path. Move the Current view to its own
/attendance/current route and turn /attendance into a query-preserving
redirect, keeping each tab on a distinct URL.

Signed kiosk URLs (/attendance?mode=kiosk&sig=...&ts=...&nonce=...) keep
working: the redirect carries the query string through. Nav, badge, and
search-index entries still point at /attendance and follow the redirect.
